safechat-ml-backend/requirements.txt
fastapi
uvicorn
transformers[torch]
transformers
torch
accelerate
huggingface-hub
websockets
httpx
python-dotenv
loguru
orjson
pytest
pytest-asyncio
